cloud/requirements.txt
functions-framework==2.2.1
google-cloud-storage==1.41.0
pg8000==1.20.0
SQLAlchemy==1.4.21
requests~=2.25.1